+++ CHANGELOG.md
@@ -19,6 +19,9 @@ Breaking changes:
 Other notable changes:
 * `webapp-util`:
   * Fixed `Rotator` and `Saver` to not bother "preserving" empty files.
+* `util`:
+  * New class `BaseValueVisitor` which implements the "visitor" pattern to
+    iterate over object graphs.
 
 ### v0.8.1 -- 2024-09-26